separate the auth routes from the user routes
