+++ backend/routes/user.go
@@ -8,9 +8,6 @@ import (
 )
 
 func UserRoutes(router *httprouter.Router) {
-	router.POST("/login", controllers.Login)
-	router.GET("/auth", middlewares.Authenticate(controllers.AuthenticateUser))
-
 	router.GET("/user", middlewares.Authenticate(controllers.ListUser))
 	router.GET("/user/:id", middlewares.Authenticate(controllers.GetUser))
 	router.POST("/user", middlewares.Authenticate(controllers.CreateUser))
